+++ src/templates/share.js
@@ -2,6 +2,6 @@
  * src/templates/share.js
  * Share page template - delegates to base HTML wrapper
  */
-import { HTML } from './base'
+import { HTML } from './base.js'
 
 export const Share = data => HTML(data)